Update the README file to have better grammar, and more useful info
This updates the `README.md` file to improve the grammar and phrasing of certain
sentences, as well as to be more precise or provide clarifications where needed.

This also adds a section to point to the origin of this
repo (`github.com/satori/go.uuid`) and mentions why the fork happened.

This also introduces a new License and Go Report Card badge at the top of the
file.

Signed-off-by: Tim Heckman <t@heckman.io>