Disable testing against Go 1.3 ~ 1.6 to improve tests
In preparation for executing on #15 (already opened in #17) we need to disable
testing against older versions of the Go toolchain. There are a few reasons for
this, the largest being they don't support subtests. These versions, including
Go 1.7 and 1.8, are EOL by the Go authors so there's also lack of upstream
support for these toolchains.

This was discussed a bit in #18 and on [Slack](https://gophers.slack.com/archives/CBP4N9BEU/p1531704009000014).

Fixes #18

Signed-off-by: Tim Heckman <t@heckman.io>