Deprecate the Equal package function
This deprecates the `Equal` package function, in favor of consumers using the
`==` comparison operator themselves between two `UUID` values. Because the UUID
value is an array type, there is no reason for the helper function.

The intent is to remove this as part of a v2.x => v3.x release.

Fixes #35

Signed-off-by: Tim Heckman <t@heckman.io>