Expose the reference UUID generator constructor
This change exposes the reference UUID generator in the package, so that
consumers can create a new instance and make use of it. This is being done so
that consumers can provide their own Hardware Address generation function, in
case they want to randomize the MAC address as to not expose it.

This works by now exposing the `rfc4122Generator` type as `Gen`, and adding
constructor functions that scaffold the generator appropriately. The tests were
also enhanced to confirm that the `HWAddrFunc` we provide is indeed being
invoked.

Fixes #41

Signed-off-by: Tim Heckman <t@heckman.io>