move code to v3 subdirectory, add forwarding root package
In #61, @bcmills describes one possible solution to the current
import path conundrum:

> Hmm, come to think of it, you might not even need to switch to major
> version 4. You could make v3.2.0 a forwarding package (without a
> go.mod) to github.com/gofrs/uuid/v3), with a corresponding v3.3.0
> (with a go.mod) that provides the /v3 module.

This patch is a demo of the first phase of this solution. It removes
the go.mod file, moves existing code to the new /v3 subdirectory,
and adds a forward.go file in the root directory, with forwarding
declarations and type aliases, targetting /v3.

For discussion and comparison with the approach in #62.