Remove testing multiple packages in coverage.
Coverprofile only supports multi-package in Go 1.10+ and therefore this breaks the CI build when we added the forwarding package. See #66 comment from @acln0. 

As there are no tests in the alias package that surfaced this issue, we aren't removing any coverage. But it is something to be aware of if another package is added to this repo.