Update V2 test to use delay to guarantee unique UUID generation
This change updates the uniqueness test for the `NewV2()` function to have a
time hop between the two generations of the UUID. This should avoid known
deficiencies in V2 UUIDs causing the test to fail. Specifically, that unique
UUIDs are only generated every 7 seconds.

This enhances the comment on the NewV2 function to make it clear that using it
is not recommended, and that it has some deficiencies.

Finally, I added some `#nosec` comments to make `gosec` not flag our usage of
SHA1 or MD5 as being unsafe.

Fixes #69

Signed-off-by: Tim Heckman <t@heckman.io>