Adds Compare and UUID.Equal funcs
Adds a Compare and UUID.Equal functions:

Compare takes two UUID's and returns -1 if the first UUID is lexically
before the second, 0 if they are equal, and 1 if the first UUID is
lexically after the second.

This implementation follows RFC 4122 section 4.5 where equality and
order checks should be done by converting the UUID to a unsigned 128-bit
integer and compared.

Tests included.